requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-mail==1.5.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.6
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
supervisor==4.2.5
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
